+++ src/components/CandyAlertState.tsx
@@ -1,0 +1,5 @@
+export default interface AlertState {
+  open: boolean;
+  message: string;
+  severity: "success" | "info" | "warning" | "error" | undefined;
+}
